refactor(settings): move settings helpers to kobo/settings/utils.py DEV-1976
